Tehty tehtävä 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -3,11 +3,14 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
+
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
+
+
 
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id).populate('user', { username: 1, name: 1 })
@@ -57,10 +60,32 @@ blogsRouter.post('/', async (request, response) => {
     response.status(201).json(savedBlog)
 })
 
+
+
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() == decodedToken.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()    
+  } else {
+    return response.status(403).json({ error: 'permission denied' })
+  }
 })
+
 
 blogsRouter.put('/:id', async (request, response) => {
   const { title, author, url, likes } = request.body
@@ -77,5 +102,6 @@ blogsRouter.put('/:id', async (request, response) => {
     response.status(404).end()
   }
 })
+
 
 module.exports = blogsRouter
